move color thumb from color picker
