+++ apps/builder/app/builder/features/style-panel/shared/color-picker.tsx
@@ -1,4 +1,11 @@
-import { useEffect, useState } from "react";
+import {
+  forwardRef,
+  useEffect,
+  useState,
+  type ComponentProps,
+  type RefObject,
+  type ElementRef,
+} from "react";
 import { colord, extend, type RgbaColor } from "colord";
 import namesPlugin from "colord/plugins/names";
 import { type ColorResult, type RGBColor, SketchPicker } from "react-color";
@@ -24,6 +31,7 @@ import { CssValueInput } from "./css-value-input";
 import type { IntermediateStyleValue } from "./css-value-input/css-value-input";
 import { isCanvasPointerEventsEnabledStore } from "~/builder/shared/nano-states";
 import { useStore } from "@nanostores/react";
+import { ColorThumb } from "./color-thumb";
 
 // To support color names
 extend([namesPlugin]);
@@ -87,37 +95,6 @@ const styleValueToRgbaColor = (value: CssColorPickerValueInput): RgbaColor => {
     b: color.b,
     a: color.a,
   };
-};
-
-const whiteColor = { r: 255, g: 255, b: 255, a: 1 };
-const borderColorSwatch = colord(rawTheme.colors.borderColorSwatch).toRgb();
-const transparent = colord("transparent").toRgb();
-
-const distance = (a: RgbaColor, b: RgbaColor) => {
-  // Use Euclidian distance, if will not give good results use https://zschuessler.github.io/DeltaE/
-  return Math.sqrt(
-    Math.pow(a.r / 255 - b.r / 255, 2) +
-      Math.pow(a.g / 255 - b.g / 255, 2) +
-      Math.pow(a.b / 255 - b.b / 255, 2) +
-      Math.pow(a.a - b.a, 2)
-  );
-};
-
-const lerp = (a: number, b: number, t: number) => {
-  return a * (1 - t) + b * t;
-};
-
-const lerpColor = (a: RgbaColor, b: RgbaColor, t: number) => {
-  return {
-    r: lerp(a.r, b.r, t),
-    g: lerp(a.g, b.g, t),
-    b: lerp(a.b, b.b, t),
-    a: lerp(a.a, b.a, t),
-  };
-};
-
-const clamp = (value: number, min: number, max: number) => {
-  return Math.min(Math.max(value, min), max);
 };
 
 // Dragging over canvas iframe with CORS policies will lead to loosing events and getting stuck in mousedown state.
@@ -212,34 +189,6 @@ export const ColorPicker = ({
     };
   }
 
-  const prefixColorRgba = styleValueToRgbaColor(prefixColor);
-
-  // @todo transparent icon can be better
-  const background =
-    prefixColorRgba.a < 1
-      ? // chessboard 5x5
-        `repeating-conic-gradient(rgba(0,0,0,0.22) 0% 25%, transparent 0% 50%) 0% 33.33% / 40% 40%, ${toValue(
-          prefixColor
-        )}`
-      : toValue(prefixColor);
-
-  const distanceToStartDrawBorder = 0.15;
-
-  // White color is invisible on white background, so we need to draw border
-  // the more color is white the more border is visible
-  const borderColor = colord(
-    lerpColor(
-      transparent,
-      borderColorSwatch,
-      clamp(
-        (distanceToStartDrawBorder - distance(whiteColor, prefixColorRgba)) /
-          distanceToStartDrawBorder,
-        0,
-        1
-      )
-    )
-  ).toRgbString();
-
   /**
    * By default, the color can be transparent, but if the user chooses a color from the picker,
    * we must set alpha = 1 otherwise all selected colors will be transparent.
@@ -267,18 +216,7 @@ export const ColorPicker = ({
         aria-label="Open color picker"
         onClick={() => setDisplayColorPicker((shown) => !shown)}
       >
-        <Box
-          css={{
-            margin: theme.spacing[3],
-            width: theme.spacing[10],
-            height: theme.spacing[10],
-            backgroundBlendMode: "difference",
-            borderRadius: 2,
-            background,
-            borderColor,
-            borderStyle: "solid",
-          }}
-        />
+        <ColorThumb color={rgbValue} />
       </PopoverTrigger>
 
       <PopoverContent>
